backend/requirements.txt
# FastAPI and server
fastapi~=0.122.0
uvicorn[standard]~=0.38.0
python-multipart~=0.0.20

# Database
sqlalchemy~=2.0.44
alembic~=1.17.2
psycopg2-binary~=2.9.11

# Validation
pydantic~=2.12.5
pydantic[email]
pydantic-settings~=2.12.0

# Authentication
python-jose[cryptography]~=3.5.0
bcrypt~=4.0.1
auth0-python~=4.13.0

# AI/ML
langchain~=1.1.0
langchain-openai~=1.1.0
langgraph~=1.0.4
deepagents~=0.2.8
openai~=2.8.1
sentence-transformers~=5.1.2

# Vector DB
qdrant-client~=1.16.1
langchain-qdrant~=1.1.0

# Platform integrations
google-auth~=2.35.0
google-auth-oauthlib~=1.2.1
google-auth-httplib2~=0.2.0
google-api-python-client~=2.149.0
slack-sdk~=3.26.1

# Utilities
python-dotenv~=1.2.1
httpx~=0.28.1
tenacity~=9.1.2

# Testing
pytest~=9.0.1
pytest-asyncio~=1.3.0
pytest-cov~=7.0.0
hypothesis~=6.148.3
